+++ src/FacebookAds/ApiConfig.php
@@ -24,6 +24,6 @@
 namespace FacebookAds;
 class ApiConfig {
   const APIVersion = '8.0';
-  const SDKVersion = '8.0.4';
+  const SDKVersion = '9.0.0';
   const TYPE_CHECKER_STRICT_MODE = false;
 }