+++ src/FacebookAds/Object/Fields/AdgroupActivityFields.php
@@ -81,6 +81,8 @@ class AdgroupActivityFields extends AbstractEnum {
   const STOP_TIME_OLD = 'stop_time_old';
   const TARGET_SPEC_ID_NEW = 'target_spec_id_new';
   const TARGET_SPEC_ID_OLD = 'target_spec_id_old';
+  const TARGETS_SPEC_NEW = 'targets_spec_new';
+  const TARGETS_SPEC_OLD = 'targets_spec_old';
   const TRACKING_PIXEL_IDS_NEW = 'tracking_pixel_ids_new';
   const TRACKING_PIXEL_IDS_OLD = 'tracking_pixel_ids_old';
   const TRACKING_SPECS_NEW = 'tracking_specs_new';
@@ -136,6 +138,8 @@ class AdgroupActivityFields extends AbstractEnum {
       'stop_time_old' => 'datetime',
       'target_spec_id_new' => 'string',
       'target_spec_id_old' => 'string',
+      'targets_spec_new' => 'Object',
+      'targets_spec_old' => 'Object',
       'tracking_pixel_ids_new' => 'list<string>',
       'tracking_pixel_ids_old' => 'list<string>',
       'tracking_specs_new' => 'list<Object>',