+++ src/FacebookAds/Object/Post.php
@@ -367,6 +367,7 @@ class Post extends AbstractCrudObject {
         'SHOP_NOW',
         'SIGN_UP',
         'SOTTO_SUBSCRIBE',
+        'START_ORDER',
         'SUBSCRIBE',
         'SWIPE_UP_PRODUCT',
         'SWIPE_UP_SHOP',