+++ src/FacebookAds/Object/Values/CampaignObjectiveValues.php
@@ -46,6 +46,7 @@ class CampaignObjectiveValues extends AbstractEnum {
   const LOCAL_AWARENESS = 'LOCAL_AWARENESS';
   const MESSAGES = 'MESSAGES';
   const OFFER_CLAIMS = 'OFFER_CLAIMS';
+  const OUTCOME_LEADS = 'OUTCOME_LEADS';
   const PAGE_LIKES = 'PAGE_LIKES';
   const POST_ENGAGEMENT = 'POST_ENGAGEMENT';
   const PRODUCT_CATALOG_SALES = 'PRODUCT_CATALOG_SALES';