+++ src/FacebookAds/Object/Values/CampaignSmartPromotionTypeValues.php
@@ -33,23 +33,10 @@ use FacebookAds\Enum\AbstractEnum;
  * on github and we'll fix in our codegen framework. We'll not be able to accept
  * pull request for this class.
  *
- * @method static SystemUserRoleValues getInstance()
+ * @method static CampaignSmartPromotionTypeValues getInstance()
  */
-class SystemUserRoleValues extends AbstractEnum {
+class CampaignSmartPromotionTypeValues extends AbstractEnum {
 
-  const ADMIN = 'ADMIN';
-  const ADS_RIGHTS_REVIEWER = 'ADS_RIGHTS_REVIEWER';
-  const VALUE_DEFAULT = 'DEFAULT';
-  const DEVELOPER = 'DEVELOPER';
-  const EMPLOYEE = 'EMPLOYEE';
-  const FINANCE_ANALYST = 'FINANCE_ANALYST';
-  const FINANCE_EDIT = 'FINANCE_EDIT';
-  const FINANCE_EDITOR = 'FINANCE_EDITOR';
-  const FINANCE_VIEW = 'FINANCE_VIEW';
-  const MANAGE = 'MANAGE';
-  const PARTNER_CENTER_ADMIN = 'PARTNER_CENTER_ADMIN';
-  const PARTNER_CENTER_ANALYST = 'PARTNER_CENTER_ANALYST';
-  const PARTNER_CENTER_EDUCATION = 'PARTNER_CENTER_EDUCATION';
-  const PARTNER_CENTER_MARKETING = 'PARTNER_CENTER_MARKETING';
-  const PARTNER_CENTER_OPERATIONS = 'PARTNER_CENTER_OPERATIONS';
+  const GUIDED_CREATION = 'GUIDED_CREATION';
+  const SMART_APP_PROMOTION = 'SMART_APP_PROMOTION';
 }